[SMALL FIX] add more ctors in BlockMasterClientPool and FileSystemMasterClientPool to help testing

+++ core/common/src/main/java/alluxio/Constants.java
@@ -52,7 +52,7 @@ public final class Constants {
   public static final String MESOS_RESOURCE_DISK = "disk";
   public static final String MESOS_RESOURCE_PORTS = "ports";
 
-  public static final int SECOND_NANO = 1000000000;
+  public static final long SECOND_NANO = 1000000000L;
   public static final int SECOND_MS = 1000;
   public static final int MINUTE_MS = SECOND_MS * 60;
   public static final int HOUR_MS = MINUTE_MS * 60;
